Move biometric signatures into plugin
