+++ Core/Lib/AjaxForms/SalesController.php
@@ -26,7 +26,6 @@ use FacturaScripts\Core\Lib\ExtendedController\BaseView;
 use FacturaScripts\Core\Lib\ExtendedController\DocFilesTrait;
 use FacturaScripts\Core\Lib\ExtendedController\LogAuditTrait;
 use FacturaScripts\Core\Lib\ExtendedController\PanelController;
-use FacturaScripts\Core\Session;
 use FacturaScripts\Core\Model\Base\SalesDocument;
 use FacturaScripts\Core\Model\Base\SalesDocumentLine;
 use FacturaScripts\Core\Tools;
@@ -34,7 +33,6 @@ use FacturaScripts\Dinamic\Lib\AssetManager;
 use FacturaScripts\Dinamic\Model\Cliente;
 use FacturaScripts\Dinamic\Model\RoleAccess;
 use FacturaScripts\Dinamic\Model\Variante;
-use FacturaScripts\Dinamic\Model\DocumentSignature;
 
 /**
  * Description of SalesController
@@ -142,14 +140,8 @@ abstract class SalesController extends PanelController
     {
         $this->setTabsPosition('top');
         $this->createViewsDoc();
-        $this->createViewSignatures();
         $this->createViewDocFiles();
         $this->createViewLogAudit();
-    }
-
-    protected function createViewSignatures(string $viewName = 'signatures'): void
-    {
-        $this->addHtmlView($viewName, 'Tab/DocumentSignatures', 'DocumentSignature', 'signature', 'fa-solid fa-signature');
     }
 
     protected function createViewsDoc(): void
@@ -200,12 +192,6 @@ abstract class SalesController extends PanelController
 
             case 'autocomplete-product':
                 return $this->autocompleteProductAction();
-
-            case 'save-signature':
-                return $this->saveSignatureAction();
-
-            case 'delete-signature':
-                return $this->deleteSignatureAction();
 
             case 'add-product':
             case 'fast-line':
@@ -328,10 +314,6 @@ abstract class SalesController extends PanelController
                 $this->loadDataDocFiles($view, $this->getModelClassName(), $code);
                 break;
 
-            case 'signatures':
-                $this->loadDataDocumentSignatures($view, $this->getModelClassName(), $code);
-                break;
-
             case 'ListLogMessage':
                 $this->loadDataLogAudit($view, $this->getModelClassName(), $code);
                 break;
@@ -360,165 +342,6 @@ abstract class SalesController extends PanelController
                 ]);
                 break;
         }
-    }
-
-    private function loadDataDocumentSignatures($view, string $model, string $code): void
-    {
-        if (empty($code)) {
-            return;
-        }
-
-        $where = [new DataBaseWhere('model', $model)];
-        if (is_numeric($code)) {
-            $where[] = new DataBaseWhere('modelid|modelcode', $code);
-        } else {
-            $where[] = new DataBaseWhere('modelcode', $code);
-        }
-
-        $view->loadData('', $where, ['signed_at' => 'DESC']);
-    }
-
-    private function saveSignatureAction(): bool
-    {
-        if (false === $this->permissions->allowUpdate) {
-            Tools::log()->warning('not-allowed-modify');
-            return true;
-        } elseif (false === $this->validateFormToken()) {
-            return true;
-        }
-
-        $model = $this->getModel();
-        if (empty($model->id())) {
-            Tools::log()->warning('record-not-found');
-            return true;
-        }
-
-        $signerName = trim((string)$this->request->input('signer_name'));
-        $signerEmail = trim((string)$this->request->input('signer_email'));
-        $signerPhone = trim((string)$this->request->input('signer_phone'));
-        $signerId = trim((string)$this->request->input('signer_id'));
-        $signerNotes = trim((string)$this->request->input('signer_notes'));
-        $signatureImage = (string)$this->request->input('signature_image');
-
-        if (empty($signerName)) {
-            Tools::log()->warning('field-required', ['%field%' => Tools::trans('name')]);
-            return true;
-        }
-
-        if (!empty($signerEmail) && false === filter_var($signerEmail, FILTER_VALIDATE_EMAIL)) {
-            Tools::log()->warning('not-valid-email', ['%email%' => $signerEmail]);
-            return true;
-        }
-
-        if (empty($signatureImage)) {
-            Tools::log()->warning('field-required', ['%field%' => Tools::trans('signature')]);
-            return true;
-        }
-
-        if (strpos($signatureImage, ',') !== false) {
-            $signatureImage = explode(',', $signatureImage, 2)[1];
-        }
-
-        $binarySignature = base64_decode($signatureImage);
-        if (false === $binarySignature || empty($binarySignature)) {
-            Tools::log()->error('record-save-error');
-            return true;
-        }
-
-        $relativeFolder = 'MyFiles/Signatures/' . date('Y') . '/' . date('m');
-        $folderPath = FS_FOLDER . '/' . $relativeFolder;
-        if (false === Tools::folderCheckOrCreate($folderPath)) {
-            Tools::log()->critical('cant-create-folder', ['%folderName%' => $relativeFolder]);
-            return true;
-        }
-
-        $documentCode = $model->codigo ?? (string)$model->id();
-        $normalizedCode = strtolower(preg_replace('/[^a-z0-9_-]+/i', '-', $documentCode));
-        $normalizedCode = trim($normalizedCode, '-');
-        if (empty($normalizedCode)) {
-            $normalizedCode = (string)$model->id();
-        }
-
-        $fileName = strtolower($this->getModelClassName()) . '-' . $normalizedCode
-            . '-' . date('YmdHis') . '-' . Tools::randomString(6) . '.png';
-        $fullPath = $folderPath . '/' . $fileName;
-
-        if (false === file_put_contents($fullPath, $binarySignature)) {
-            Tools::log()->error('record-save-error');
-            return true;
-        }
-
-        $relativePath = $relativeFolder . '/' . $fileName;
-
-        $signature = new DocumentSignature();
-        $signature->model = $this->getModelClassName();
-        $signature->modelcode = $documentCode;
-        $signature->modelid = (int)$model->id();
-        $signature->signer_name = $signerName;
-        $signature->signer_email = $signerEmail;
-        $signature->signer_phone = $signerPhone;
-        $signature->signer_id = $signerId;
-        $signature->signer_notes = $signerNotes;
-        $signature->signature_path = $relativePath;
-        $signature->signed_at = Tools::dateTime();
-        $signature->ip_address = (string)Session::getClientIp();
-        $userAgent = (string)$this->request->headers->get('User-Agent');
-        $signature->user_agent = substr($userAgent, 0, 255);
-        $signature->nick = $this->user->nick;
-
-        if ($signature->save()) {
-            Tools::log()->notice('record-updated-correctly');
-            return true;
-        }
-
-        if (file_exists($fullPath)) {
-            @unlink($fullPath);
-        }
-
-        Tools::log()->error('record-save-error');
-        return true;
-    }
-
-    private function deleteSignatureAction(): bool
-    {
-        if (false === $this->permissions->allowDelete) {
-            Tools::log()->warning('not-allowed-delete');
-            return true;
-        } elseif (false === $this->validateFormToken()) {
-            return true;
-        }
-
-        $signatureId = (int)$this->request->input('id');
-        $signature = new DocumentSignature();
-        if (false === $signature->load($signatureId)) {
-            Tools::log()->warning('record-not-found');
-            return true;
-        }
-
-        $model = $this->getModel();
-        if (empty($model->id())) {
-            Tools::log()->warning('record-not-found');
-            return true;
-        }
-
-        $modelId = (int)$model->id();
-        $modelCode = $model->codigo ?? (string)$modelId;
-
-        $matchesId = $signature->modelid === $modelId;
-        $matchesCode = in_array($signature->modelcode, [$modelCode, (string)$modelId], true);
-
-        if ($signature->model !== $this->getModelClassName() || (!$matchesId && !$matchesCode)) {
-            Tools::log()->warning('not-allowed-delete');
-            return true;
-        }
-
-        if ($signature->delete()) {
-            Tools::log()->notice('record-deleted-correctly');
-            return true;
-        }
-
-        Tools::log()->warning('record-deleted-error');
-        return true;
     }
 
     protected function recalculateAction(bool $renderLines): bool
