+++ Core/Translation/ca_ES.json
@@ -1470,8 +1470,6 @@
     "show-price": "Mostra preu en la impressió",
     "show-quantity": "Mostra quantitat en la impressió",
     "show-section-new": "Mostra la secció crear",
-    "signature-instructions": "Signa amb el dit o amb el ratolí i prem Desa per adjuntar la signatura juntament amb les dades del signant.",
-    "signature-requires-saved-document": "Desa el document abans de capturar una signatura.",
     "signature": "Signatura",
     "simple": "Simple",
     "simplified": "Simplificada",
