+++ Core/Translation/cs_CZ.json
@@ -1470,8 +1470,6 @@
     "show-price": "Mostrar precio en la impresión",
     "show-quantity": "Mostrar cantidad en la impresión",
     "show-section-new": "Zobrazit sekci pro vytvoření",
-    "signature-instructions": "Sign using your mouse or finger, then press Save to attach the signature together with the signer details.",
-    "signature-requires-saved-document": "Save the document before capturing a signature.",
     "signature": "Firma",
     "simple": "Simple",
     "simplified": "Zjednodušený",
