+++ Core/Translation/en_EN.json
@@ -1470,8 +1470,6 @@
     "show-price": "Show price on print",
     "show-quantity": "Show quantity on print",
     "show-section-new": "Show create section",
-    "signature-instructions": "Sign using your mouse or finger, then press Save to attach the signature together with the signer details.",
-    "signature-requires-saved-document": "Save the document before capturing a signature.",
     "signature": "Signature",
     "simple": "Simple",
     "simplified": "Simplified",
