+++ Core/Translation/fr_FR.json
@@ -1470,8 +1470,6 @@
     "show-price": "Afficher le prix sur l'impression",
     "show-quantity": "Afficher la quantité sur l'impression",
     "show-section-new": "Afficher la section de création",
-    "signature-instructions": "Signez avec votre doigt ou la souris, puis cliquez sur Enregistrer pour joindre la signature et les données du signataire.",
-    "signature-requires-saved-document": "Enregistrez le document avant de capturer une signature.",
     "signature": "Signature",
     "simple": "Simple",
     "simplified": "Simplifié",
