+++ Core/Translation/gl_ES.json
@@ -1470,8 +1470,6 @@
     "show-price": "Mostrar prezo na impresión",
     "show-quantity": "Mostrar cantidade na impresión",
     "show-section-new": "Mostrar sección crear",
-    "signature-instructions": "Asina co dedo ou co rato e preme Gardar para anexar a sinatura xunto cos datos da persoa asinante.",
-    "signature-requires-saved-document": "Garda o documento antes de capturar unha sinatura.",
     "signature": "Sinatura",
     "simple": "Simple",
     "simplified": "Simplificada",
