+++ Core/Translation/it_IT.json
@@ -1470,8 +1470,6 @@
     "show-price": "Mostra prezzo su stampa",
     "show-quantity": "Mostra la quantità sulla stampa",
     "show-section-new": "Mostra sezione creazione",
-    "signature-instructions": "Firma con il dito o con il mouse e premi Salva per allegare la firma insieme ai dati del firmatario.",
-    "signature-requires-saved-document": "Salva il documento prima di acquisire una firma.",
     "signature": "Firma",
     "simple": "Semplice",
     "simplified": "semplificata",
