+++ Core/Translation/pl_PL.json
@@ -1470,8 +1470,6 @@
     "show-price": "Pokaż cenę w druku",
     "show-quantity": "Pokaż ilość w druku",
     "show-section-new": "Pokaż sekcję tworzenia",
-    "signature-instructions": "Sign using your mouse or finger, then press Save to attach the signature together with the signer details.",
-    "signature-requires-saved-document": "Save the document before capturing a signature.",
     "signature": "Podpis",
     "simple": "Prosty",
     "simplified": "Uproszczona",
