+++ Core/Translation/pt_BR.json
@@ -1470,8 +1470,6 @@
     "show-price": "Mostrar preço na impressão",
     "show-quantity": "Mostrar quantidade ao imprimir",
     "show-section-new": "Mostrar seção nova",
-    "signature-instructions": "Assine com o dedo ou com o mouse e clique em Salvar para anexar a assinatura junto com os dados do signatário.",
-    "signature-requires-saved-document": "Salve o documento antes de capturar uma assinatura.",
     "signature": "Assinatura",
     "simple": "Simples",
     "simplified": "Simplificado",
