+++ Core/Translation/pt_PT.json
@@ -1470,8 +1470,6 @@
     "show-price": "Mostrar preço na impressão",
     "show-quantity": "Mostrar quantidade ao imprimir",
     "show-section-new": "Mostrar seção nova",
-    "signature-instructions": "Assine com o dedo ou com o rato e clique em Guardar para anexar a assinatura juntamente com os dados do signatário.",
-    "signature-requires-saved-document": "Guarde o documento antes de capturar uma assinatura.",
     "signature": "Assinatura",
     "simple": "Simples",
     "simplified": "Simplificada",
