--- Dinamic/Model/DocumentSignature.php
@@ -1,8 +1,0 @@
-<?php namespace FacturaScripts\Dinamic\Model;
-
-/**
- * Class created to extend the core DocumentSignature model.
- */
-class DocumentSignature extends \FacturaScripts\Core\Model\DocumentSignature
-{
-}
