+++ Plugins/BiometricSignature/Model/DocumentSignature.php
@@ -17,7 +17,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace FacturaScripts\Core\Model;
+namespace FacturaScripts\Plugins\BiometricSignature\Model;
 
 use FacturaScripts\Core\Lib\MyFilesToken;
 use FacturaScripts\Core\Template\ModelClass;
