+++ Plugins/BiometricSignature/Table/document_signatures.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
     Document   : document_signatures.xml
-    Author     : ChatGPT
+    Author     : BiometricSignature plugin
     Description:
         Structure for the document_signatures table.
 -->
