+++ Plugins/BiometricSignature/Translation/en_EN.json
@@ -1,0 +1,4 @@
+{
+    "signature-instructions": "Sign using your mouse or finger, then press Save to attach the signature together with the signer details.",
+    "signature-requires-saved-document": "Save the document before capturing a signature."
+}
