+++ Plugins/BiometricSignature/Translation/es_ES.json
@@ -1,0 +1,4 @@
+{
+    "signature-instructions": "Firma con el dedo o el ratón y pulsa Guardar para adjuntar la firma junto con los datos del firmante.",
+    "signature-requires-saved-document": "Guarda el documento antes de capturar la firma."
+}
