+++ Plugins/BiometricSignature/facturascripts.ini
@@ -1,0 +1,4 @@
+name = 'BiometricSignature'
+description = 'Firma biométrica de documentos de venta con datos del firmante.'
+version = 1
+min_version = 2025
